Patrol Directions
Patrol directions can be inscribed into the roomEnemyMapData.map file.

Since the file is inverted on the y axis, up is 'V' and down is '^' in code, but in the file it looks normal.

For example in:
^ x x
T  x T
x  x V

The first T from the left is going down, but the second one is actually going up.